Tidy file structure and improve naming
