+++ public/js/containers/OriginData.js
@@ -4,7 +4,8 @@ import {
     getInWorkflowVsNotInWorkflow
     
 } from "../selectors/charts";
-import { getFilterVals, getIsUpdating } from "../selectors";
+import { getFilterVals } from "../selectors/filters";
+import { getIsUpdating } from "../selectors/isUpdating";
 import { toggleStackChart } from "../actions";
 import Origin from "../components/Tabs/Origin";
