+++ public/js/selectors/commissionedLength.js
@@ -1,24 +1,4 @@
-import moment from "moment";
 import { createSelector } from "reselect";
-
-export const getIsUpdating = ({ isUpdating }) => isUpdating;
-
-/* Filters */
-
-const getFilters = ({ filters }) => filters;
-
-export const getFilterVals = createSelector(getFilters, ({ values }) => ({
-    ...values,
-    startDate: moment(values.startDate),
-    endDate: moment(values.endDate)
-}));
-
-export const getFilterStatuses = createSelector(
-    getFilters,
-    ({ statuses }) => statuses
-);
-
-/* Commissioned length */
 
 const getArticleWordCounts = ({ articleWordCounts }) => articleWordCounts;
